backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pyjwt
bcrypt==4.1.3
python-dotenv==1.0.1
passlib==1.7.4
pycryptodome
argon2_cffi